requirements.txt
streamlit[auth]>=1.42
gspread>=6.0
google-auth>=2.29
pandas>=2.2
Authlib>=1.3.2
